fixed main menu for mobile
