--- hamza-client/src/modules/layout/templates/nav-4/components/nav-link.tsx
@@ -1,25 +1,0 @@
-'use client';
-
-import Link from 'next/link';
-import { useParams } from 'next/navigation';
-import React from 'react';
-
-interface NavLinkProps {
-    href: string;
-    children: any;
-}
-
-const NavLink: React.FC<NavLinkProps> = ({ href, children }) => {
-    const { countryCode } = useParams();
-
-    return (
-        <Link
-            style={{ padding: '0', margin: '0' }}
-            href={`/${countryCode}${href}`}
-        >
-            {children}
-        </Link>
-    );
-};
-
-export default NavLink;
