+++ hamza-client/src/modules/layout/templates/nav/menu-desktop/components/nav-searchbar.tsx
@@ -1,12 +1,20 @@
 'use client';
-
 import React from 'react';
-import { Flex, Input, InputGroup, InputRightElement } from '@chakra-ui/react';
+import {
+    Box,
+    Flex,
+    Input,
+    InputGroup,
+    InputLeftElement,
+    InputRightElement,
+} from '@chakra-ui/react';
 import { FaSearch } from 'react-icons/fa';
 import { useEffect, useState } from 'react';
 import { usePathname, useRouter } from 'next/navigation';
 import SearchModal from '@modules/search/templates/search-modal';
 import { getRegion } from 'app/actions';
+import fslash from '../../.../../../../../../../public/images/nav/search/fslash.svg';
+import Image from 'next/image';
 
 const NavSearchBar = () => {
     const [searchOpened, setSearchOpened] = useState(false);
@@ -27,15 +35,20 @@ const NavSearchBar = () => {
     }, [pathname]);
 
     return (
-        <>
-            <InputGroup mx="2rem" display={isHomePage ? 'none' : 'block'}>
+        <Flex width={'100%'} display={{ base: 'none', md: 'flex' }}>
+            <InputGroup
+                display={isHomePage ? 'none' : 'flex'}
+                mx="1rem"
+                alignItems={'center'}
+            >
                 <Input
                     width={'100%'}
                     height={'80px'}
                     border="1px solid"
-                    borderRadius="40px"
+                    borderRadius="16px"
                     borderColor="#555555"
-                    backgroundColor="black"
+                    backgroundColor="#121212"
+                    pl="4rem"
                     fontSize={{ base: '14px', md: '16px' }}
                     color="#3E3E3E"
                     placeholder="Search for product name, product..."
@@ -47,15 +60,34 @@ const NavSearchBar = () => {
                         setSearchOpened(true);
                     }}
                 />
-                <InputRightElement
+                <InputLeftElement
                     height="100%"
-                    display="flex"
                     alignItems="center"
                     pointerEvents="none"
-                    mr="1rem"
+                    ml="1rem"
                     fontSize={{ base: '20px', md: '28px' }}
                 >
-                    <FaSearch color="#555555" />
+                    <FaSearch color="white" />
+                </InputLeftElement>
+                <InputRightElement height="100%" mr="1rem">
+                    <Flex
+                        borderRadius={'10px'}
+                        height="47px"
+                        width="45px"
+                        justifyContent={'center'}
+                        alignItems="center"
+                        alignSelf={'center'}
+                        backgroundColor={'#272727'}
+                    >
+                        <Image
+                            src={fslash}
+                            alt="slash"
+                            style={{
+                                width: '11.39px',
+                                height: 'auto',
+                            }}
+                        />
+                    </Flex>
                 </InputRightElement>
             </InputGroup>
             {searchOpened && (
@@ -65,7 +97,7 @@ const NavSearchBar = () => {
                     }}
                 />
             )}
-        </>
+        </Flex>
     );
 };
