+++ hamza-client/src/modules/layout/templates/nav/menu-mobile/mobile-nav.tsx
@@ -4,8 +4,8 @@ import { Flex, Text } from '@chakra-ui/react';
 import Image from 'next/image';
 import HamzaLogo from '../../../../../../public/images/logo/logo_green.svg';
 import HamzaTitle from '../../../../../../public/images/logo/hamza-title.svg';
-import MobileMenu from '../menu/mobile-menu';
-import ConnectWallet from '../connect-mobile';
+import MobileMenu from './mobile-menu';
+import ConnectWallet from './connect-mobile';
 
 // import CartButton from '@modules/layout/components/cart-button';
 import CartButton from '@modules/layout/components/cart-button-mobile';
